ensure legends are independent of one another
